разнес все по файлам, оставил только program...
